Include more parts of the request in traces
So far our traces for requests have not included any of the HTTP
headers. The reason for that was that the only headers that we set are
for the key-id and secret key, and we want to avoid including them in
traces. However, with recent changes we may end up including an
Accepted-Encoding header as well. It would be good to be able to infer
its state from traces.
With this change we make that happen, by including more of the request
meta data in traces. Right now that's the HTTP version in use and the
headers. However, we make sure to mask the contents of the key-id and
secret headers -- if users are really in need of debugging them, they
should be able to poke at the processes environment directly.