Rename data::v2::stream::Normalized to SymbolList
This change renames the data::v2::stream::Normalized type to SymbolList.
The name Normalized was always seen as a bit thorny, as it had this aura
of being an implementation detail. It is not, and the hope is that with
the new name the intention and its status is conveyed more clearly.