Rework symbols related types in data::v2::stream
The representation we chose for a list of symbols to subscribe to turned
out to be not quite optimal. We were enforcing the property that if a
Symbol::All variant is present, then it has to be the only symbol at
runtime, while that could be compile-time enforced. That representation
originated in another crate, which had slightly different constraints
and where that made sense, but we missed to rework the types for this
crate's use case.
With this change we overhaul the types, such that we now have a
top-level enum, Symbols, that contains an All variant or a list of
symbols -- making it impossible to violate the aforementioned
constraint.